+++ src/services/permanentStorage/cookiePermanentStorage.service.ts
@@ -1,7 +1,9 @@
-import {Injectable} from '@angular/core';
+import {ClassProvider, Injectable, Provider} from '@angular/core';
 
-import {PermanentStorage} from './permanentStorage.interface';
+import {PermanentStorage, PermanentStorageType} from './permanentStorage.interface';
 import {CookieService} from '../cookies/cookies.service';
+import {PERMANENT_STORAGE} from '../../types/tokens';
+import {TypeProvider} from '../../types/providerDecoratedType';
 
 /**
  * Implementation of permanent storage using cookies
@@ -60,3 +62,27 @@ export class CookiePermanentStorageService implements PermanentStorage
         this._cookies.deleteCookie(name, '/');
     }
 }
+
+
+/**
+ * Provider for permanent storage that is using cookie storage implementation
+ */
+const COOKIE_PERMANENT_STORAGE: Provider =
+<ClassProvider>
+{
+    provide: PERMANENT_STORAGE,
+    useClass: CookiePermanentStorageService
+};
+
+/**
+ * Type that contains provider for cookie permanent storage when used with `providePermanentStorage`
+ */
+@TypeProvider(COOKIE_PERMANENT_STORAGE)
+class CookiePermanentStorageType
+{
+}
+
+/**
+ * Sets permanent storage to use cookie permanent storage when used with `providePermanentStorage`
+ */
+export const CookiePermanentStorage: PermanentStorageType = CookiePermanentStorageType;