+++ src/services/stringLocalization/stringLocalization.interface.ts
@@ -1,3 +1,4 @@
+import {Type} from '@angular/core';
 import {Observable} from 'rxjs';
 
 /**
@@ -16,4 +17,11 @@ export interface StringLocalization
      * @param interpolateParams - Optional object storing interpolation parameters
      */
     get(key: string, interpolateParams?: Object): string;
+}
+
+/**
+ * Used for restriction of string localization provider type only for type decorated with string localization provider
+ */
+export interface StringLocalizationType extends Type<unknown>
+{
 }