+++ store/src/services/permanentStorage/localPermanentStorage.service.ts
@@ -1,9 +1,7 @@
-import {Injectable} from '@angular/core';
-import {PermanentStorage} from '@anglr/common';
+import {ClassProvider, Injectable, Provider} from '@angular/core';
+import {PERMANENT_STORAGE, PermanentStorage, PermanentStorageType, TypeProvider} from '@anglr/common';
 import store from 'store';
 import expirePlugin from 'store/plugins/expire';
-
-//TODO: add tokens for storages
 
 /**
  * Implementation of permanent storage using LocalStorage
@@ -71,3 +69,26 @@ export class LocalPermanentStorageService implements PermanentStorage
         store.remove(name);
     }
 }
+
+/**
+ * Provider for permanent storage that is using local storage implementation
+ */
+const LOCAL_PERMANENT_STORAGE: Provider =
+<ClassProvider>
+{
+    provide: PERMANENT_STORAGE,
+    useClass: LocalPermanentStorageService
+};
+
+/**
+ * Type that contains provider for local permanent storage when used with `providePermanentStorage`
+ */
+@TypeProvider(LOCAL_PERMANENT_STORAGE)
+class LocalPermanentStorageType
+{
+}
+
+/**
+ * Sets permanent storage to use local permanent storage when used with `providePermanentStorage`
+ */
+export const LocalPermanentStorage: PermanentStorageType = LocalPermanentStorageType;