+++ structured-log/src/index.ts
@@ -1,6 +1,7 @@
 export * from './components/console/console.component.animations';
 export * from './components/console/console.component';
 export * from './modules/consoleLog.module';
+export * from './misc/providers';
 export * from './misc/utils';
 export * from './types/tokens';
 export * from './types/logger.interface';